# apps/web/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "DemoCorpusAttachmentResponse": {
        "properties": {
          "byte_length": {
            "title": "Byte Length",
            "type": "integer"
          },
          "evidence_id": {
            "format": "uuid",
            "title": "Evidence Id",
            "type": "string"
          },
          "media_type": {
            "title": "Media Type",
            "type": "string"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          }
        },
        "required": [
          "evidence_id",
          "media_type",
          "byte_length",
          "sha256"
        ],
        "title": "DemoCorpusAttachmentResponse",
        "type": "object"
      },
      "DemoCorpusMessageResponse": {
        "description": "One message in exactly the shape `POST /ingest/messages` accepts for a single item.\n\nSo the frontend never needs to reshape this response before replaying it \u2014 see\n`DemoCorpusResponse` (P2-7).",
        "properties": {
          "adapter": {
            "const": "SYNTHETIC",
            "title": "Adapter",
            "type": "string"
          },
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/DemoCorpusAttachmentResponse"
            },
            "title": "Attachments",
            "type": "array"
          },
          "channel_message_id": {
            "title": "Channel Message Id",
            "type": "string"
          },
          "contributor_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contributor Id"
          },
          "sent_at": {
            "title": "Sent At",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "adapter",
          "channel_message_id",
          "contributor_id",
          "sent_at",
          "text",
          "attachments"
        ],
        "title": "DemoCorpusMessageResponse",
        "type": "object"
      },
      "DemoCorpusResponse": {
        "description": "The immutable public ingestion input for the demo corpus, server-owned (P2-7).\n\nBefore this route, the only way a browser could replay the seeded corpus to trigger the\nMonitor was to mirror the checked-in fixture file as a frontend constant -- a second copy\nof `demo/fixtures/elevator-v1/{manifest,feed}.json` that could silently drift from the one\nthe server actually seeds with. This route serves the same\n:class:`~chorus.infrastructure.fixtures.synthetic_feed.SyntheticAmbientAdapter` reset\nitself reads, so a fixture change on the server is picked up by the browser with no\nfrontend change at all.\n\nIt carries no secret: every field here is exactly what `GET /feed` already shows a\npresenter after reset, resolved into the shape `POST /ingest/messages` accepts. It is\npresenter-only and wired only where `reset_demo` is (local/demo compositions), because it\nexists to serve the demo's own replay step and nothing else.",
        "properties": {
          "community_id": {
            "format": "uuid",
            "title": "Community Id",
            "type": "string"
          },
          "corpus_sha256": {
            "title": "Corpus Sha256",
            "type": "string"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/DemoCorpusMessageResponse"
            },
            "title": "Messages",
            "type": "array"
          },
          "seed_version": {
            "title": "Seed Version",
            "type": "string"
          }
        },
        "required": [
          "seed_version",
          "corpus_sha256",
          "community_id",
          "messages"
        ],
        "title": "DemoCorpusResponse",
        "type": "object"
      },
...
      "FeedItemResponse": {
        "properties": {
          "attachment_thumbnails": {
            "items": {
              "$ref": "#/components/schemas/AttachmentThumbnailResponse"
            },
            "title": "Attachment Thumbnails",
            "type": "array"
          },
          "chorus_signal": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChorusSignalResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message_id": {
            "format": "uuid",
            "title": "Message Id",
            "type": "string"
          },
          "pseudonym": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pseudonym"
          },
          "sent_at": {
            "format": "date-time",
            "title": "Sent At",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "message_id",
          "sent_at",
          "pseudonym",
          "text",
          "attachment_thumbnails",
          "chorus_signal"
        ],
        "title": "FeedItemResponse",
        "type": "object"
      },
      "FeedResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FeedItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "FeedResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IngestMessageRequest": {
        "additionalProperties": false,
        "properties": {
          "adapter": {
            "const": "SYNTHETIC",
            "default": "SYNTHETIC",
            "title": "Adapter",
            "type": "string"
          },
          "attachments": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/IngestAttachmentRequest"
            },
            "maxItems": 8,
            "title": "Attachments",
            "type": "array"
          },
          "channel_message_id": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Channel Message Id",
            "type": "string"
          },
          "contributor_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contributor Id"
          },
          "sent_at": {
            "format": "date-time",
            "title": "Sent At",
            "type": "string"
          },
          "text": {
            "maxLength": 10000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "channel_message_id",
          "sent_at",
          "text"
        ],
        "title": "IngestMessageRequest",
        "type": "object"
      },
      "IngestMessagesRequest": {
        "additionalProperties": false,
        "properties": {
          "community_id": {
            "format": "uuid",
            "title": "Community Id",
            "type": "string"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/IngestMessageRequest"
            },
            "maxItems": 25,
            "minItems": 1,
            "title": "Messages",
            "type": "array"
          }
        },
        "required": [
          "community_id",
          "messages"
        ],
        "title": "IngestMessagesRequest",
        "type": "object"
      },
      "IngestMessagesResponse": {
        "properties": {
          "accepted_count": {
            "title": "Accepted Count",
            "type": "integer"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/IngestedMessageResponse"
            },
            "title": "Messages",
            "type": "array"
          },
          "operation": {
            "$ref": "#/components/schemas/OperationReference"
          },
          "replayed_count": {
            "title": "Replayed Count",
            "type": "integer"
          }
        },
        "required": [
          "messages",
          "accepted_count",
          "replayed_count",
          "operation"
        ],
        "title": "IngestMessagesResponse",
        "type": "object"
      },
      "IngestedMessageResponse": {
        "properties": {
          "channel_message_id": {
            "title": "Channel Message Id",
            "type": "string"
          },
          "message_id": {
            "format": "uuid",
            "title": "Message Id",
            "type": "string"
          },
          "replay": {
            "title": "Replay",
            "type": "boolean"
          }
        },
        "required": [
          "channel_message_id",
          "message_id",
          "replay"
        ],
        "title": "IngestedMessageResponse",
        "type": "object"
      },
...
      "OperationReference": {
        "properties": {
          "operation_id": {
            "format": "uuid",
            "title": "Operation Id",
            "type": "string"
          },
          "poll_url": {
            "title": "Poll Url",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "operation_id",
          "status",
          "poll_url"
        ],
        "title": "OperationReference",
        "type": "object"
      },
...
      "ResetContributorResponse": {
        "properties": {
          "actor": {
            "title": "Actor",
            "type": "string"
          },
          "contributor_id": {
            "format": "uuid",
            "title": "Contributor Id",
            "type": "string"
          },
          "pseudonym": {
            "title": "Pseudonym",
            "type": "string"
          }
        },
        "required": [
          "actor",
          "pseudonym",
          "contributor_id"
        ],
        "title": "ResetContributorResponse",
        "type": "object"
      },
      "ResetCountsResponse": {
        "properties": {
          "contributors": {
            "title": "Contributors",
            "type": "integer"
          },
          "deleted": {
            "title": "Deleted",
            "type": "integer"
          },
          "evidence": {
            "title": "Evidence",
            "type": "integer"
          },
          "messages": {
            "title": "Messages",
            "type": "integer"
          }
        },
        "required": [
          "deleted",
          "messages",
          "contributors",
          "evidence"
        ],
        "title": "ResetCountsResponse",
        "type": "object"
      },
      "ResetEvidenceResponse": {
        "properties": {
          "evidence_id": {
            "format": "uuid",
            "title": "Evidence Id",
            "type": "string"
          },
          "media_type": {
            "title": "Media Type",
            "type": "string"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          }
        },
        "required": [
          "evidence_id",
          "media_type",
          "sha256"
        ],
        "title": "ResetEvidenceResponse",
        "type": "object"
      },
      "ResetRequest": {
        "additionalProperties": false,
        "properties": {
          "confirm": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Confirm",
            "type": "string"
          },
          "namespace": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Namespace",
            "type": "string"
          },
          "seed_version": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Seed Version",
            "type": "string"
          }
        },
        "required": [
          "namespace",
          "confirm",
          "seed_version"
        ],
        "title": "ResetRequest",
        "type": "object"
      },
      "ResetResponse": {
        "properties": {
          "audit_event_id": {
            "format": "uuid",
            "title": "Audit Event Id",
            "type": "string"
          },
          "community_id": {
            "format": "uuid",
            "title": "Community Id",
            "type": "string"
          },
          "contributors": {
            "items": {
              "$ref": "#/components/schemas/ResetContributorResponse"
            },
            "title": "Contributors",
            "type": "array"
          },
          "corpus_sha256": {
            "title": "Corpus Sha256",
            "type": "string"
          },
          "counts": {
            "$ref": "#/components/schemas/ResetCountsResponse"
          },
          "destination_id": {
            "title": "Destination Id",
            "type": "string"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/ResetEvidenceResponse"
            },
            "title": "Evidence",
            "type": "array"
          },
          "logical_now": {
            "title": "Logical Now",
            "type": "string"
          },
          "namespace": {
            "title": "Namespace",
            "type": "string"
          },
          "replayed": {
            "title": "Replayed",
            "type": "boolean"
          },
          "reset_id": {
            "format": "uuid",
            "title": "Reset Id",
            "type": "string"
          },
          "seed_version": {
            "title": "Seed Version",
            "type": "string"
          }
        },
        "required": [
          "reset_id",
          "namespace",
          "seed_version",
          "corpus_sha256",
          "logical_now",
          "community_id",
          "destination_id",
          "contributors",
          "evidence",
          "counts",
          "replayed",
          "audit_event_id"
        ],
        "title": "ResetResponse",
        "type": "object"
      },
...
      "SessionResponse": {
        "properties": {
          "actor": {
            "title": "Actor",
            "type": "string"
          },
          "capabilities": {
            "items": {
              "type": "string"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "community_id": {
            "format": "uuid",
            "title": "Community Id",
            "type": "string"
          },
          "contributor_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contributor Id"
          },
          "namespace": {
            "title": "Namespace",
            "type": "string"
          }
        },
        "required": [
          "actor",
          "contributor_id",
          "community_id",
          "namespace",
          "capabilities"
        ],
        "title": "SessionResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Ambient CHORUS",
    "version": "0.9.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/demo/corpus": {
      "get": {
        "description": "Serve the exact seeded corpus, resolved into the shape ingest already accepts.",
        "operationId": "read_demo_corpus_v1_demo_corpus_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Chorus-Demo-Actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chorus-Demo-Actor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemoCorpusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Demo Corpus",
        "tags": [
          "demo"
        ]
      }
    },
...
    "/v1/demo/reset": {
      "post": {
        "description": "Delete-and-seed the demo namespace, idempotently, against local adapters.",
        "operationId": "reset_demo_v1_demo_reset_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "minLength": 8,
                  "pattern": "^[\\x20-\\x7e]+$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "X-Chorus-Demo-Actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chorus-Demo-Actor"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reset Demo",
        "tags": [
          "demo"
        ]
      }
    },
    "/v1/feed": {
      "get": {
        "description": "Return one page of ambient messages with any discovered-pattern signal attached.",
        "operationId": "read_feed_v1_feed_get",
        "parameters": [
          {
            "in": "query",
            "name": "community_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Community Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 2048,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "header",
            "name": "X-Chorus-Demo-Actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chorus-Demo-Actor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Feed",
        "tags": [
          "feed"
        ]
      }
    },
    "/v1/ingest/messages": {
      "post": {
        "description": "Persist a batch of ambient messages and start one Monitor operation over it.",
        "operationId": "ingest_messages_v1_ingest_messages_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 8,
              "pattern": "^[\\x20-\\x7e]+$",
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Chorus-Demo-Actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chorus-Demo-Actor"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestMessagesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestMessagesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Messages",
        "tags": [
          "ingest"
        ]
      }
    },
...
    "/v1/session": {
      "get": {
        "description": "Resolve the calling persona. Cannot enumerate or name any other persona.",
        "operationId": "read_session_v1_session_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Chorus-Demo-Actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chorus-Demo-Actor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Session",
        "tags": [
          "demo"
        ]
      }
    }
  }
}